# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "GRIDT API Documentation",
    "description": "this is a description of the API that will be written at some point"
  },
  "servers": [{ "url": "http://localhost:3333/api/v1" }],
...
  "paths": {
...
    "/schools": {
      "get": {
        "tags": ["school"],
        "summary": "Find all Schools",
        "description": "Returns all School documents in schools collection of GRIDT",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["school"],
        "summary": "Add one new School",
        "description": "Add one School document in schools collection of GRIDT",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/School"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Success" },
          "400": { "description": "Bad Request" },
          "409": {
            "description": "COMBOKEY property must be unique"
          },
          "422": {
            "description": "name, COMBOKEY, and SCH_NAME properties are required"
          }
        }
      },
      "put": {
        "tags": ["school"],
        "summary": "Update one School by ID",
        "description": "Updates one School document by ID in schools collection of GRIDT",
        "parameters": [{ "in": "path", "name": "id" }],
        "responses": {
          "204": { "description": "Success" },
          "400": { "description": "Bad request" },
          "500": { "description": "Server error" }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/School"
              }
            }
          }
        }
      }
    },
    "/schools/{id}": {
      "get": {
        "tags": ["school"],
        "summary": "Find one School by ID",
        "description": "Returns one School document by ID in schools collection of GRIDT",
        "responses": { "200": { "description": "OK" } },
        "parameters": [{ "in": "path", "name": "id" }]
      },
      "delete": {
        "tags": ["school"],
        "summary": "Delete one School by ID",
        "description": "Deletes one School document by ID in schools collection of GRIDT",
        "responses": {
          "202": { "description": "Success" },
          "400": { "description": "Bad request" },
          "404": { "description": "Item not found" }
        },
        "parameters": [{ "in": "path", "name": "id" }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "School": {
        "title": "School",
        "type": "object",
        "required": [
          "LEA_STATE",
          "LEA_STATE_NAME",
          "LEAID",
          "LEA_NAME",
          "SCHID",
          "SCH_NAME",
          "COMBOKEY",
          "SCH_IDEAENR_HI_M",
          "SCH_ENR_HI_M",
          "RI_HI_M",
          "TOT_IDEA_W/O_HI_M",
          "TOT_ENR_W/O_HI_M",
          "RI_W/O_HI_M",
          "RR_HI_M_POP",
          "RR_HI_M_WH_M",
          "SCH_IDEAENR_HI_F",
          "SCH_ENR_HI_F",
          "RI_HI_F",
          "TOT_IDEA_W/O_HI_F",
          "TOT_ENR_W/O_HI_F",
          "RI_W/O_HI_F",
          "RR_HI_F_POP",
          "RR_HI_F_WH_F",
          "SCH_IDEAENR_AM_M",
          "SCH_ENR_AM_M",
          "RI_AM_M",
          "TOT_IDEA_W/O_AM_M",
          "TOT_ENR_W/O_AM_M",
          "RI_W/O_AM_M",
          "RR_AM_M_POP",
          "RR_AM_M_WH_M",
          "SCH_IDEAENR_AM_F",
          "SCH_ENR_AM_F",
          "RI_AM_F",
          "TOT_IDEA_W/O_AM_F",
          "TOT_ENR_W/O_AM_F",
          "RI_W/O_AM_F",
          "RR_AM_F_POP",
          "RR_AM_F_WH_F",
          "SCH_IDEAENR_AS_M",
          "SCH_ENR_AS_M",
          "RI_AS_M",
          "TOT_IDEA_W/O_AS_M",
          "TOT_ENR_W/O_AS_M",
          "RI_W/O_AS_M",
          "RR_AS_M_POP",
          "RR_AS_M_WH_M",
          "SCH_IDEAENR_AS_F",
          "SCH_ENR_AS_F",
          "RI_AS_F",
          "TOT_IDEA_W/O_AS_F",
          "TOT_ENR_W/O_AS_F",
          "RI_W/O_AS_F",
          "RR_AS_F_POP",
          "RR_AS_F_WH_F",
          "SCH_IDEAENR_HP_M",
          "SCH_ENR_HP_M",
          "RI_HP_M",
          "TOT_IDEA_W/O_HP_M",
          "TOT_ENR_W/O_HP_M",
          "RI_W/O_HP_M",
          "RR_HP_M_POP",
          "RR_HP_M_WH_M",
          "SCH_IDEAENR_HP_F",
          "SCH_ENR_HP_F",
          "RI_HP_F",
          "TOT_IDEA_W/O_HP_F",
          "TOT_ENR_W/O_HP_F",
          "RI_W/O_HP_F",
          "RR_HP_F_POP",
          "RR_HP_F_WH_F",
          "SCH_IDEAENR_BL_M",
          "SCH_ENR_BL_M",
          "RI_BL_M",
          "TOT_IDEA_W/O_BL_M",
          "TOT_ENR_W/O_BL_M",
          "RI_W/O_BL_M",
          "RR_BL_M_POP",
          "RR_BL_M_WH_M",
          "SCH_IDEAENR_BL_F",
          "SCH_ENR_BL_F",
          "RI_BL_F",
          "TOT_IDEA_W/O_BL_F",
          "TOT_ENR_W/O_BL_F",
          "RI_W/O_BL_F",
          "RR_BL_F_POP",
          "RR_BL_F_WH_F",
          "SCH_IDEAENR_TR_M",
          "SCH_ENR_TR_M",
          "RI_TR_M",
          "TOT_IDEA_W/O_TR_M",
          "TOT_ENR_W/O_TR_M",
          "RI_W/O_TR_M",
          "RR_TR_M_POP",
          "RR_TR_M_WH_M",
          "SCH_IDEAENR_TR_F",
          "SCH_ENR_TR_F",
          "RI_TR_F",
          "TOT_IDEA_W/O_TR_F",
          "TOT_ENR_W/O_TR_F",
          "RI_W/O_TR_F",
          "RR_TR_F_POP",
          "RR_TR_F_WH_F",
          "SCH_IDEAENR_WH_M",
          "SCH_ENR_WH_M",
          "TOT_IDEA_W/O_WH_M",
          "TOT_ENR_W/O_WH_M",
          "RI_W/O_WH_M",
          "RR_WH_M_POP",
          "RR_WH_M_WH_M",
          "SCH_IDEAENR_WH_F",
          "SCH_ENR_WH_F",
          "TOT_IDEA_W/O_WH_F",
          "TOT_ENR_W/O_WH_F",
          "RI_W/O_WH_F",
          "RR_WH_F_POP",
          "RR_WH_F_WH_F",
          "TOT_IDEAENR_M",
          "TOT_IDEAENR_F",
          "TOT_IDEA_ENR",
          "TOT_ENR_M",
          "TOT_ENR_F",
          "TOT_ENR",
          "SCH_IDEA_WH_TOT",
          "SCH_ENR_WH_TOT",
          "RI_WH_TOT",
          "TOT_IDEA_W/O_WH_TOT",
          "TOT_ENR_W/O_WH_TOT",
          "RI_W/O_WH_TOT",
          "RR_WH_TOT_POP",
          "RR_WH_TOT_WH_TOT",
          "RR_HI_M_WH_TOT",
          "RR_HI_F_WH_TOT",
          "RR_AM_M_WH_TOT",
          "RR_AM_F_WH_TOT",
          "RR_AS_M_WH_TOT",
          "RR_AS_F_WH_TOT",
          "RR_HP_M_WH_TOT",
          "RR_HP_F_WH_TOT",
          "RR_BL_M_WH_TOT",
          "RR_BL_F_WH_TOT",
          "RR_TR_M_WH_TOT",
          "RR_TR_F_WH_TOT",
          "RI_WH_M",
          "RI_WH_F"
        ],
        "properties": {
          "LEA_STATE": { "type": "string" },
          "LEA_STATE_NAME": { "type": "string" },
          "LEAID": { "type": "string" },
          "LEA_NAME": { "type": "string" },
          "SCHID": { "type": "string" },
          "SCH_NAME": { "type": "string" },
          "COMBOKEY": { "type": "number" },
          "SCH_IDEAENR_HI_M": { "type": "string" },
          "SCH_ENR_HI_M": { "type": "string" },
          "RI_HI_M": { "type": "string" },
          "TOT_IDEA_W/O_HI_M": { "type": "string" },
          "TOT_ENR_W/O_HI_M": { "type": "string" },
          "RI_W/O_HI_M": { "type": "string" },
          "RR_HI_M_POP": { "type": "string" },
          "RR_HI_M_WH_M": { "type": "string" },
          "SCH_IDEAENR_HI_F": { "type": "string" },
          "SCH_ENR_HI_F": { "type": "string" },
          "RI_HI_F": { "type": "string" },
          "TOT_IDEA_W/O_HI_F": { "type": "string" },
          "TOT_ENR_W/O_HI_F": { "type": "string" },
          "RI_W/O_HI_F": { "type": "string" },
          "RR_HI_F_POP": { "type": "string" },
          "RR_HI_F_WH_F": { "type": "string" },
          "SCH_IDEAENR_AM_M": { "type": "string" },
          "SCH_ENR_AM_M": { "type": "string" },
          "RI_AM_M": { "type": "string" },
          "TOT_IDEA_W/O_AM_M": { "type": "string" },
          "TOT_ENR_W/O_AM_M": { "type": "string" },
          "RI_W/O_AM_M": { "type": "string" },
          "RR_AM_M_POP": { "type": "string" },
          "RR_AM_M_WH_M": { "type": "string" },
          "SCH_IDEAENR_AM_F": { "type": "string" },
          "SCH_ENR_AM_F": { "type": "string" },
          "RI_AM_F": { "type": "string" },
          "TOT_IDEA_W/O_AM_F": { "type": "string" },
          "TOT_ENR_W/O_AM_F": { "type": "string" },
          "RI_W/O_AM_F": { "type": "string" },
          "RR_AM_F_POP": { "type": "string" },
          "RR_AM_F_WH_F": { "type": "string" },
          "SCH_IDEAENR_AS_M": { "type": "string" },
          "SCH_ENR_AS_M": { "type": "string" },
          "RI_AS_M": { "type": "string" },
          "TOT_IDEA_W/O_AS_M": { "type": "string" },
          "TOT_ENR_W/O_AS_M": { "type": "string" },
          "RI_W/O_AS_M": { "type": "string" },
          "RR_AS_M_POP": { "type": "string" },
          "RR_AS_M_WH_M": { "type": "string" },
          "SCH_IDEAENR_AS_F": { "type": "string" },
          "SCH_ENR_AS_F": { "type": "string" },
          "RI_AS_F": { "type": "string" },
          "TOT_IDEA_W/O_AS_F": { "type": "string" },
          "TOT_ENR_W/O_AS_F": { "type": "string" },
          "RI_W/O_AS_F": { "type": "string" },
          "RR_AS_F_POP": { "type": "string" },
          "RR_AS_F_WH_F": { "type": "string" },
          "SCH_IDEAENR_HP_M": { "type": "string" },
          "SCH_ENR_HP_M": { "type": "string" },
          "RI_HP_M": { "type": "string" },
          "TOT_IDEA_W/O_HP_M": { "type": "string" },
          "TOT_ENR_W/O_HP_M": { "type": "string" },
          "RI_W/O_HP_M": { "type": "string" },
          "RR_HP_M_POP": { "type": "string" },
          "RR_HP_M_WH_M": { "type": "string" },
          "SCH_IDEAENR_HP_F": { "type": "string" },
          "SCH_ENR_HP_F": { "type": "string" },
          "RI_HP_F": { "type": "string" },
          "TOT_IDEA_W/O_HP_F": { "type": "string" },
          "TOT_ENR_W/O_HP_F": { "type": "string" },
          "RI_W/O_HP_F": { "type": "string" },
          "RR_HP_F_POP": { "type": "string" },
          "RR_HP_F_WH_F": { "type": "string" },
          "SCH_IDEAENR_BL_M": { "type": "string" },
          "SCH_ENR_BL_M": { "type": "string" },
          "RI_BL_M": { "type": "string" },
          "TOT_IDEA_W/O_BL_M": { "type": "string" },
          "TOT_ENR_W/O_BL_M": { "type": "string" },
          "RI_W/O_BL_M": { "type": "string" },
          "RR_BL_M_POP": { "type": "string" },
          "RR_BL_M_WH_M": { "type": "string" },
          "SCH_IDEAENR_BL_F": { "type": "string" },
          "SCH_ENR_BL_F": { "type": "string" },
          "RI_BL_F": { "type": "string" },
          "TOT_IDEA_W/O_BL_F": { "type": "string" },
          "TOT_ENR_W/O_BL_F": { "type": "string" },
          "RI_W/O_BL_F": { "type": "string" },
          "RR_BL_F_POP": { "type": "string" },
          "RR_BL_F_WH_F": { "type": "string" },
          "SCH_IDEAENR_TR_M": { "type": "string" },
          "SCH_ENR_TR_M": { "type": "string" },
          "RI_TR_M": { "type": "string" },
          "TOT_IDEA_W/O_TR_M": { "type": "string" },
          "TOT_ENR_W/O_TR_M": { "type": "string" },
          "RI_W/O_TR_M": { "type": "string" },
          "RR_TR_M_POP": { "type": "string" },
          "RR_TR_M_WH_M": { "type": "string" },
          "SCH_IDEAENR_TR_F": { "type": "string" },
          "SCH_ENR_TR_F": { "type": "string" },
          "RI_TR_F": { "type": "string" },
          "TOT_IDEA_W/O_TR_F": { "type": "string" },
          "TOT_ENR_W/O_TR_F": { "type": "string" },
          "RI_W/O_TR_F": { "type": "string" },
          "RR_TR_F_POP": { "type": "string" },
          "RR_TR_F_WH_F": { "type": "string" },
          "SCH_IDEAENR_WH_M": { "type": "string" },
          "SCH_ENR_WH_M": { "type": "string" },
          "TOT_IDEA_W/O_WH_M": { "type": "string" },
          "TOT_ENR_W/O_WH_M": { "type": "string" },
          "RI_W/O_WH_M": { "type": "string" },
          "RR_WH_M_POP": { "type": "string" },
          "RR_WH_M_WH_M": { "type": "string" },
          "SCH_IDEAENR_WH_F": { "type": "string" },
          "SCH_ENR_WH_F": { "type": "string" },
          "TOT_IDEA_W/O_WH_F": { "type": "string" },
          "TOT_ENR_W/O_WH_F": { "type": "string" },
          "RI_W/O_WH_F": { "type": "string" },
          "RR_WH_F_POP": { "type": "string" },
          "RR_WH_F_WH_F": { "type": "string" },
          "TOT_IDEAENR_M": { "type": "string" },
          "TOT_IDEAENR_F": { "type": "string" },
          "TOT_IDEA_ENR": { "type": "string" },
          "TOT_ENR_M": { "type": "string" },
          "TOT_ENR_F": { "type": "string" },
          "TOT_ENR": { "type": "string" },
          "SCH_IDEA_WH_TOT": { "type": "string" },
          "SCH_ENR_WH_TOT": { "type": "string" },
          "RI_WH_TOT": { "type": "string" },
          "TOT_IDEA_W/O_WH_TOT": { "type": "string" },
          "TOT_ENR_W/O_WH_TOT": { "type": "string" },
          "RI_W/O_WH_TOT": { "type": "string" },
          "RR_WH_TOT_POP": { "type": "string" },
          "RR_WH_TOT_WH_TOT": { "type": "string" },
          "RR_HI_M_WH_TOT": { "type": "string" },
          "RR_HI_F_WH_TOT": { "type": "string" },
          "RR_AM_M_WH_TOT": { "type": "string" },
          "RR_AM_F_WH_TOT": { "type": "string" },
          "RR_AS_M_WH_TOT": { "type": "string" },
          "RR_AS_F_WH_TOT": { "type": "string" },
          "RR_HP_M_WH_TOT": { "type": "string" },
          "RR_HP_F_WH_TOT": { "type": "string" },
          "RR_BL_M_WH_TOT": { "type": "string" },
          "RR_BL_F_WH_TOT": { "type": "string" },
          "RR_TR_M_WH_TOT": { "type": "string" },
          "RR_TR_F_WH_TOT": { "type": "string" },
          "RI_WH_M": { "type": "string" },
          "RI_WH_F": { "type": "string" }
        }
      }
    }
  }
}