# server/resources/schools/school.model.js
const mongoose = require('mongoose');
const { composeMongoose } = require('graphql-compose-mongoose');

const schoolSchema = new mongoose.Schema({
  LEA_STATE: { type: String },
  LEA_STATE_NAME: { type: String },
  LEAID: { type: Number },
  LEA_NAME: { type: String },
  SCHID: { type: Number },
  SCH_NAME: { type: String },
  COMBOKEY: { type: String },
  SCH_IDEAENR_HI_M: { type: Number },
  SCH_ENR_HI_M: { type: Number },
  RI_HI_M: { type: Number },
  TOT_IDEA_WO_HI_M: { type: Number },
  TOT_ENR_WO_HI_M: { type: Number },
  RI_WO_HI_M: { type: Number },
  RR_HI_M_POP: { type: Number },
  RR_HI_M_WH_M: { type: Number },
  SCH_IDEAENR_HI_F: { type: Number },
  SCH_ENR_HI_F: { type: Number },
  RI_HI_F: { type: Number },
  TOT_IDEA_WO_HI_F: { type: Number },
  TOT_ENR_WO_HI_F: { type: Number },
  RI_WO_HI_F: { type: Number },
  RR_HI_F_POP: { type: Number },
  RR_HI_F_WH_F: { type: Number },
  SCH_IDEAENR_AM_M: { type: Number },
  SCH_ENR_AM_M: { type: Number },
  RI_AM_M: { type: Number },
  TOT_IDEA_WO_AM_M: { type: Number },
  TOT_ENR_WO_AM_M: { type: Number },
  RI_WO_AM_M: { type: Number },
  RR_AM_M_POP: { type: Number },
  RR_AM_M_WH_M: { type: Number },
  SCH_IDEAENR_AM_F: { type: Number },
  SCH_ENR_AM_F: { type: Number },
  RI_AM_F: { type: Number },
  TOT_IDEA_WO_AM_F: { type: Number },
  TOT_ENR_WO_AM_F: { type: Number },
  RI_WO_AM_F: { type: Number },
  RR_AM_F_POP: { type: Number },
  RR_AM_F_WH_F: { type: Number },
  SCH_IDEAENR_AS_M: { type: Number },
  SCH_ENR_AS_M: { type: Number },
  RI_AS_M: { type: Number },
  TOT_IDEA_WO_AS_M: { type: Number },
  TOT_ENR_WO_AS_M: { type: Number },
  RI_WO_AS_M: { type: Number },
  RR_AS_M_POP: { type: Number },
  RR_AS_M_WH_M: { type: Number },
  SCH_IDEAENR_AS_F: { type: Number },
  SCH_ENR_AS_F: { type: Number },
  RI_AS_F: { type: Number },
  TOT_IDEA_WO_AS_F: { type: Number },
  TOT_ENR_WO_AS_F: { type: Number },
  RI_WO_AS_F: { type: Number },
  RR_AS_F_POP: { type: Number },
  RR_AS_F_WH_F: { type: Number },
  SCH_IDEAENR_HP_M: { type: Number },
  SCH_ENR_HP_M: { type: Number },
  RI_HP_M: { type: Number },
  TOT_IDEA_WO_HP_M: { type: Number },
  TOT_ENR_WO_HP_M: { type: Number },
  RI_WO_HP_M: { type: Number },
  RR_HP_M_POP: { type: Number },
  RR_HP_M_WH_M: { type: Number },
  SCH_IDEAENR_HP_F: { type: Number },
  SCH_ENR_HP_F: { type: Number },
  RI_HP_F: { type: Number },
  TOT_IDEA_WO_HP_F: { type: Number },
  TOT_ENR_WO_HP_F: { type: Number },
  RI_WO_HP_F: { type: Number },
  RR_HP_F_POP: { type: Number },
  RR_HP_F_WH_F: { type: Number },
  SCH_IDEAENR_BL_M: { type: Number },
  SCH_ENR_BL_M: { type: Number },
  RI_BL_M: { type: Number },
  TOT_IDEA_WO_BL_M: { type: Number },
  TOT_ENR_WO_BL_M: { type: Number },
  RI_WO_BL_M: { type: Number },
  RR_BL_M_POP: { type: Number },
  RR_BL_M_WH_M: { type: Number },
  SCH_IDEAENR_BL_F: { type: Number },
  SCH_ENR_BL_F: { type: Number },
  RI_BL_F: { type: Number },
  TOT_IDEA_WO_BL_F: { type: Number },
  TOT_ENR_WO_BL_F: { type: Number },
  RI_WO_BL_F: { type: Number },
  RR_BL_F_POP: { type: Number },
  RR_BL_F_WH_F: { type: Number },
  SCH_IDEAENR_TR_M: { type: Number },
  SCH_ENR_TR_M: { type: Number },
  RI_TR_M: { type: Number },
  TOT_IDEA_WO_TR_M: { type: Number },
  TOT_ENR_WO_TR_M: { type: Number },
  RI_WO_TR_M: { type: Number },
  RR_TR_M_POP: { type: Number },
  RR_TR_M_WH_M: { type: Number },
  SCH_IDEAENR_TR_F: { type: Number },
  SCH_ENR_TR_F: { type: Number },
  RI_TR_F: { type: Number },
  TOT_IDEA_WO_TR_F: { type: Number },
  TOT_ENR_WO_TR_F: { type: Number },
  RI_WO_TR_F: { type: Number },
  RR_TR_F_POP: { type: Number },
  RR_TR_F_WH_F: { type: Number },
  SCH_IDEAENR_WH_M: { type: Number },
  SCH_ENR_WH_M: { type: Number },
  TOT_IDEA_WO_WH_M: { type: Number },
  TOT_ENR_WO_WH_M: { type: Number },
  RI_WO_WH_M: { type: Number },
  RR_WH_M_POP: { type: Number },
  RR_WH_M_WH_M: { type: Number },
  SCH_IDEAENR_WH_F: { type: Number },
  SCH_ENR_WH_F: { type: Number },
  TOT_IDEA_WO_WH_F: { type: Number },
  TOT_ENR_WO_WH_F: { type: Number },
  RI_WO_WH_F: { type: Number },
  RR_WH_F_POP: { type: Number },
  RR_WH_F_WH_F: { type: Number },
  TOT_IDEAENR_M: { type: Number },
  TOT_IDEAENR_F: { type: Number },
  TOT_IDEA_ENR: { type: Number },
  TOT_ENR_M: { type: Number },
  TOT_ENR_F: { type: Number },
  TOT_ENR: { type: Number },
  SCH_IDEA_WH_TOT: { type: Number },
  SCH_ENR_WH_TOT: { type: Number },
  RI_WH_TOT: { type: Number },
  TOT_IDEA_WO_WH_TOT: { type: Number },
  TOT_ENR_WO_WH_TOT: { type: Number },
  RI_WO_WH_TOT: { type: Number },
  RR_WH_TOT_POP: { type: Number },
  RR_WH_TOT_WH_TOT: { type: Number },
  RR_HI_M_WH_TOT: { type: Number },
  RR_HI_F_WH_TOT: { type: Number },
  RR_AM_M_WH_TOT: { type: Number },
  RR_AM_F_WH_TOT: { type: Number },
  RR_AS_M_WH_TOT: { type: Number },
  RR_AS_F_WH_TOT: { type: Number },
  RR_HP_M_WH_TOT: { type: Number },
  RR_HP_F_WH_TOT: { type: Number },
  RR_BL_M_WH_TOT: { type: Number },
  RR_BL_F_WH_TOT: { type: Number },
  RR_TR_M_WH_TOT: { type: Number },
  RR_TR_F_WH_TOT: { type: Number },
  RI_WH_M: { type: Number },
  RI_WH_F: { type: Number },
});

const School = mongoose.model('School', schoolSchema);
const SchoolTC = composeMongoose(School);

module.exports = { School, SchoolTC };